+++ src/base-element.js
@@ -5,4 +5,5 @@ import {
 import { DefineableMixin } from "web-component-define";
 import { LitScopedElementsMixin } from "web-component-define/exports/lit.js"
 
+/** @type HTMLElement */
 export class BaseElement extends DefineableMixin(LitScopedElementsMixin(LitElement)) {}