+++ src/index.d.ts
@@ -14,7 +14,7 @@ export interface INinjaAction {
   id?: string;
   title?: string;
   hotkey?: string;
-  handler?: Function;
+  handler?: (action: NinjaAction, event: CustomEvent<NinjaAction>) => void;
   mdIcon?: string;
   icon?: string;
   parent?: string;
@@ -22,12 +22,37 @@ export interface INinjaAction {
   children?: string[];
   section?: string;
   content?: string;
+
+  /** When an href is defined, render an anchor tag. */
+  href?: string;
+  attributes?: Omit<Anchorable, "href">;
 }
 
-// declare global {
-//   interface HTMLElementTagNameMap {
-//     'ninja-keys': NinjaKeys;
-//     'ninja-header': NinjaHeader;
-//     'ninja-action': NinjaAction;
-//   }
-// }
+type Prettify<T> = {
+  [K in keyof T]: T[K]
+} & {}
+
+type X = Prettify<INinjaAction["attributes"]>
+
+/**
+ * Attributes for anchors, not exhaustive, but ones that matter.
+ * https://developer.mozilla.org/en-US/docs/Web/HTML/Element/a
+ */
+export interface Anchorable {
+  download?: string;
+  href?: string;
+  hreflang?: string;
+  ping?: string;
+  referrerpolicy?: string;
+  rel?: string;
+  target?: string;
+  type?: string;
+}
+
+declare global {
+  interface HTMLElementTagNameMap {
+    'ninja-keys': HTMLElement & NinjaKeys;
+    'ninja-header': HTMLElement & NinjaHeader;
+    'ninja-action': HTMLElement & NinjaAction;
+  }
+}