+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.6.0
 
 - feat: add block content to allow for more than just a title to be displayed.
+- feat: allow rendering of anchor tags that don't need handlers.
 
 ## 1.5.1